Manual fixes for naming conflicts
This commit manually fixes name conflicts that were not automatically resolved by the code generator.

With the exception of the `message` operation, these naming conflicts arise because of multiple parameters with the same name. For example, the `update_counterexample` operation has a `text` path parameter and a `text` body variable. The generator tries to “explode” the body model into the function signature, but fails if names conflict, as they do here.

To manually resolve this problem, I updated the code to append `new_` to the troublesome body variables. We may later add support for this pattern to the code generator.

The `message` endpoint required manual handling because of the `input` variable. With naive code generation, the `input` variable will shadow the built-in `input` function. So we rename it here to `message_input` instead.